% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame_table1.R
\name{as_data_frame_table1}
\alias{as_data_frame_table1}
\title{Convert table1 into data.frame}
\usage{
as_data_frame_table1(x, ...)
}
\arguments{
\item{x}{table1 object}

\item{...}{additional arguments (see table1 documentation)}
}
\description{
Convert table1 into data.frame
}
