% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate_weights.R
\name{ate_weights}
\alias{ate_weights}
\title{Calculate average treatment effect using inverse propensity score weighting}
\usage{
ate_weights(data, vars, prop.var)
}
\arguments{
\item{data}{Character vector specifying rownames of the table (empty columns should be named with "").}

\item{vars}{variables tested for Influence on outcome.}

\item{prop.var}{variable for which propensity scores should be calculated}
}
\description{
Propensity scores (PS) are calculated using fitted values obtained in a logistic regression. The inverse propensity score weighting is subsequently conducted
by 1/PS for the factor level that corresponds to 1 and by 1/(1-PS) for the factor level that corresponds to 0.
}
