% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{breakFun}
\alias{breakFun}
\title{Part of helper functions used in this package}
\usage{
breakFun(y)
}
\arguments{
\item{y}{Vector used to construct the sequence between 0 and the maximum value. If used as \code{scale_y_continuous(breaks = breakFun)},
y is inherited from the ggplot object.}
}
\description{
Creates breaks for \code{scale_y_continuous} between 0 and the maximum value of y for graphs with positive y, rounded with the \code{roundUp}
function. This ensures that the space between y-axis breaks is always a multiple of five.
}
\seealso{
\code{\link[=breakFun2]{breakFun2()}} for negative y.
}
