% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{breakFun2}
\alias{breakFun2}
\title{Part of helper functions used in this package}
\usage{
breakFun2(y)
}
\arguments{
\item{y}{Vector used to construct the sequence between 0 and the minimum value. If used as \code{scale_y_continuous(breaks = breakFun)},
y is inherited from the ggplot object.}
}
\description{
Creates breaks for \code{scale_y_continuous} between 0 and the minimum value of y for graphs with negative y, rounded with the \code{roundUp}
function. This ensures that the space between y-axis breaks is always a multiple of five.
}
\seealso{
\code{\link[=breakFun]{breakFun()}} for positive y.
}
