% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvr.R
\name{bvr}
\alias{bvr}
\title{Calculate bivariate residuals in a latent class analysis}
\usage{
bvr(fit)
}
\arguments{
\item{fit}{fit object of class "poLCA" generated with the poLCA function.}
}
\description{
This code is adapted from
Author: Daniel Oberski
Date: 2017-08-01
Bivariate residual statistic for latent class analysis
Calculate the BVR for poLCA objects
}
