% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_DCR.R
\name{calc_DCR}
\alias{calc_DCR}
\title{Calculate disease control rate}
\usage{
calc_DCR(x)
}
\arguments{
\item{x}{vector containing response data.}
}
\description{
This function takes the output of \code{convert_response} and assigns 1 if response is complete response, partial
response or stable disease and 0 if response is progressive disease.
}
