% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ORR.R
\name{calc_ORR}
\alias{calc_ORR}
\title{Calculate overall response rate}
\usage{
calc_ORR(x)
}
\arguments{
\item{x}{vector containing response data.}
}
\description{
This function takes the output of \code{convert_response} and assigns 1 if response is complete response or partial
response, otherwise assigns 0.
}
