% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_survival.R
\name{calc_survival}
\alias{calc_survival}
\title{Calculate survival interval}
\usage{
calc_survival(startDate, eventDate)
}
\arguments{
\item{startDate}{start of therapy.}

\item{eventDate}{date of event.}
}
\description{
This function calculates the survival interval between start of therapy and time of event (e.g. progression, death, etc.)
in months.
}
