% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_pval.R
\name{ci_pval}
\alias{ci_pval}
\title{Calculate pvalues from confidence intervals}
\usage{
ci_pval(est, l, u, log.trans = TRUE, conf.level = 0.95)
}
\arguments{
\item{est}{Estimate (e.g. difference or ratio)}

\item{l}{lower limit of confidence interval}

\item{u}{upper limit of confidence interval}

\item{log.trans}{logical indicating whether estimates should be log transformed. If TRUE values will be log transformed before calculation. Default is TRUE.}

\item{conf.level}{confidence level}
}
\description{
The calculation method is based on this publication: https://www.bmj.com/content/343/bmj.d2304.
}
