% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_response.R
\name{convert_response}
\alias{convert_response}
\title{Convert best clinical response to standardized format}
\usage{
convert_response(x, trtdur)
}
\arguments{
\item{x}{vector containing response data.}

\item{trtdur}{vector containing information about the treatment duration}
}
\description{
Non-assessible response is replaced by NA. NC is transformed to stable disease. NED is transformed to complete response.
Mixed response is documented as stable disease if \code{trtdur} > 90, otherwise it is documented as progressive disease.
}
