% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_output.R
\name{cox_output}
\alias{cox_output}
\title{Calculate univariate cox regression and extract Hazard Ratio and pvalues}
\usage{
cox_output(
  data,
  time,
  status,
  vars,
  modeltype = "full",
  p.thres = 0.1,
  niter = 10
)
}
\arguments{
\item{data}{data.frame or data.table containing survival data.}

\item{time}{the time interval from start of observation until date of event (e.g. disease progression or death)
or censoring.}

\item{status}{variable specifying if event occured or data has been censored.}

\item{vars}{variables tested for Influence on outcome.}

\item{modeltype}{character value. Allowed values include: "full" and "backwards".
Specify if full model or backwards selection model should be used.}

\item{p.thres}{pvalue threshold for backwards selection model.}

\item{niter}{number of iterations for backwards selection model.}
}
\description{
This function calculates a univaraite cox regression with the \code{coxph} function of the \code{survival}
package. Hazard ratios and pvalues are extracted and shown in a table format. If a variable is divided in more
than 2 groups the wald test statistic for overall significance is also shown.
}
