% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_table_combined.R
\name{cox_table_combined}
\alias{cox_table_combined}
\title{Print table with univariate, full mutlivariable and reduced (backwards selection) multivariable cox regression}
\usage{
cox_table_combined(
  data,
  time,
  status,
  vars,
  fixed.var = NULL,
  rgroup = NULL,
  p.thres = 0.1,
  footnote = NULL,
  ...
)
}
\arguments{
\item{data}{data.frame or data.table containing survival data.}

\item{time}{the time interval from start of observation until date of event (e.g. disease progression or death)
or censoring.}

\item{status}{variable specifying if event occured or data has been censored.}

\item{vars}{one or more variables defined as character strings to be included in the table}

\item{fixed.var}{specifies fixed variables to be included in the cox model.}

\item{rgroup}{a vector of character strings containing headings for row groups.}

\item{p.thres}{pvalue threshold for backwards selection model.}

\item{footnote}{character string passed on to \code{tfoot} argument of \code{htmlTable}}

\item{...}{additional arguments to be passed on to \code{cox_table_combined}}
}
\description{
This function combines the results of a univariable, full multivariable model and a backwards selection model.
}
