% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_array.R
\name{create_array}
\alias{create_array}
\title{Create an array from a list of an indexes}
\usage{
create_array(value = 1, indexes)
}
\arguments{
\item{value}{A numeric vector with the value(s) to fill the array}

\item{indexes}{A named list with the indexes of the array}
}
\value{
An array with the dimensions equal to the length of each index.
}
\description{
Create an array from a list of an indexes
}
\examples{
create_array(value = 1, indexes = list(i = c("a", "b"), j = c("c", "d")))

}
