% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_equation.R
\name{create_equation}
\alias{create_equation}
\title{Create a model equation}
\usage{
create_equation(equation, indexes = NULL, type = c("defining", "mcc"), desc)
}
\arguments{
\item{equation}{a string with the equation expression.}

\item{indexes}{a vector with index and sets specifications. Example: "c in COM".}

\item{type}{the equation type: "defining" or "mcc".}

\item{desc}{equation description.}
}
\value{
A named list with the values and type for the variable.
}
\description{
Create a model equation
}
\examples{

create_equation(
 equation = "x[c] = alpha * M / p[c]",
 indexes = "c in COM",
 type = "defining",
 desc = "Commodity demand"
)

}
