% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_param.R
\name{create_param}
\alias{create_param}
\title{Create a parameter}
\usage{
create_param(value = 1, indexes, desc)
}
\arguments{
\item{value}{a numeric vector with the value(s) to fill the array.}

\item{indexes}{a named list with the indexes of the array.}

\item{desc}{variable description.}
}
\value{
A named list with the values and type for the variable.
}
\description{
Create a parameter
}
\examples{

C <- c("sec1", "sec2", "sec3")

create_param(
 value = 1,
 indexes = list(C = C),
 desc = "Parameter description"
)

}
