% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variable.R
\name{create_variable}
\alias{create_variable}
\title{Create a variable}
\usage{
create_variable(value = 1, indexes, type = c("defined", "undefined"), desc)
}
\arguments{
\item{value}{a numeric vector with the value(s) to fill the array.}

\item{indexes}{a named list with the indexes of the array.}

\item{type}{the variable type: "defined" or "undefined".}

\item{desc}{variable description.}
}
\value{
A named list with the values and type for the variable.
}
\description{
Create a variable
}
\examples{

C <- c("sec1", "sec2", "sec3")

create_variable(
 value = 1,
 indexes = list(C = C),
 type = "defined",
 desc = "variable description"
)

}
