% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_flextable.R
\name{df_to_flextable}
\alias{df_to_flextable}
\title{Transfrom dataframe into flextable object}
\usage{
df_to_flextable(tbl, data, vars_tbl, indent = TRUE)
}
\arguments{
\item{tbl}{dataframe containing table data}

\item{data}{original data which was used to generate the table}

\item{vars_tbl}{variables used to generate the table}

\item{indent}{logical indicating whether factor levels should be indented. Default is true}
}
\description{
Transfrom dataframe into flextable object
}
