% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_cols.R
\name{emr_cols}
\alias{emr_cols}
\title{An emR function}
\usage{
emr_cols(n)
}
\arguments{
\item{n}{c("#009AA6", "#001489")}
}
\description{
This code generates a sequence between Pantone320 and Reflexblue C, the corporate design colors of EuMelaReg gGmbH.
}
