% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_describe_security_configuration}
\alias{emr_describe_security_configuration}
\title{Provides the details of a security configuration by returning the
configuration JSON}
\usage{
emr_describe_security_configuration(Name)
}
\arguments{
\item{Name}{[required] The name of the security configuration.}
}
\description{
Provides the details of a security configuration by returning the
configuration JSON.
}
\section{Request syntax}{
\preformatted{emr$describe_security_configuration(
  Name = "string"
)
}
}

\keyword{internal}
