% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_operations.R
\name{emr_remove_auto_scaling_policy}
\alias{emr_remove_auto_scaling_policy}
\title{Removes an automatic scaling policy from a specified instance group
within an EMR cluster}
\usage{
emr_remove_auto_scaling_policy(ClusterId, InstanceGroupId)
}
\arguments{
\item{ClusterId}{[required] Specifies the ID of a cluster. The instance group to which the automatic
scaling policy is applied is within this cluster.}

\item{InstanceGroupId}{[required] Specifies the ID of the instance group to which the scaling policy is
applied.}
}
\description{
Removes an automatic scaling policy from a specified instance group
within an EMR cluster.
}
\section{Request syntax}{
\preformatted{emr$remove_auto_scaling_policy(
  ClusterId = "string",
  InstanceGroupId = "string"
)
}
}

\keyword{internal}
