% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitFlextableToPage.R
\name{FitFlextableToPage}
\alias{FitFlextableToPage}
\title{Expand flextable to fit on word document}
\usage{
FitFlextableToPage(ft, pgwidth = 6)
}
\arguments{
\item{ft}{flextable object}

\item{pgwidth}{width of the desired word document. Default is 6 inches}
}
\description{
Expand flextable to fit on word document
}
