% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plotFUN.R
\name{forest_plotFUN}
\alias{forest_plotFUN}
\title{Base function to display a forestplot}
\usage{
forest_plotFUN(
  toShow,
  main,
  y_breaks,
  cpositions,
  point_size,
  fontsize,
  line_size,
  vjust_text,
  refLabel,
  noDigits,
  ylim,
  varnames
)
}
\arguments{
\item{toShow}{Results of a cox regression containing variable names, variable levels, hazards, confidence intervals and p-values, which will be converted to a forestplot.}

\item{main}{title of the plot.}

\item{y_breaks}{argument to supply manual y_breaks as a numerical vector. Default is NULL and breaks are set automatically within the function.}

\item{cpositions}{relative positions of first three columns in the OX scale.}

\item{point_size}{Size of mean points.}

\item{fontsize}{relative size of annotations in the plot. Default value: 0.7.}

\item{line_size}{Size of errorbar line.}

\item{vjust_text}{vertical adjustment of text containing information about events, global pvalue, AIC and concordance index}

\item{refLabel}{label for reference levels of factor variables.}

\item{noDigits}{number of digits for estimates and p-values in the plot.}

\item{ylim}{argument to supply manual y limits as numerical vector of length 2. Default is NULL and limits are set automatically within the function.}

\item{varnames}{Character vector specifying rownames of the table (empty columns should be named with "").}
}
\description{
This code generates a helper function used in the wrappers "forestplot_eumelareg" and "forestplot_meta_eumelareg".
}
