% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logrank_IPSW_RISCA.R
\name{logrank_IPSW_RISCA}
\alias{logrank_IPSW_RISCA}
\title{Calculate median survival time}
\usage{
logrank_IPSW_RISCA(time, status, var, weights = NULL)
}
\arguments{
\item{time}{A numeric vector with the follow up time.}

\item{status}{A numeric vector with the event indicator (0=right censored, 1=event).}

\item{var}{A numeric vector with the binary var under interest (only two groups).}

\item{weights}{The weights for correcting the contribution of each individual.
By default, the weights are all equaled to 1 and the survival curves correspond to the usual Kaplan-Meier estimator.}
}
\description{
The user enters individual survival data and the weights previously calculated (by using logistic regression for instance).
The usual log-rank test is adapted to the corresponding adjusted survival curves.
}
