% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_center_table.R
\name{make_center_table}
\alias{make_center_table}
\title{Tabulate number of patients per center and write into csv}
\usage{
make_center_table(data, filename, cent.var)
}
\arguments{
\item{data}{data.frame or data.table containing data to be tabulated.}

\item{filename}{name of the exported csv file}

\item{cent.var}{name of the center variable}
}
\description{
Tabulate number of patients per center and write into csv
}
