% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_survival_htmltable.R
\name{median_survival_htmltable}
\alias{median_survival_htmltable}
\title{Transform median survival data to htmltable format}
\usage{
median_survival_htmltable(
  data,
  timePFS = NULL,
  timeOS = NULL,
  statusPFS = NULL,
  statusOS = NULL,
  var
)
}
\arguments{
\item{data}{data.frame or data.table}

\item{timePFS}{Time for PFS}

\item{timeOS}{Time for OS}

\item{statusPFS}{censor variable for PFS}

\item{statusOS}{censor variable for OS}

\item{var}{Variable tested for Influence on outcome.}
}
\description{
Wrapper around \code{add_median_survival} to transform the data to htmltable format in order to be used in display
with \code{htmlTable::htmlTable}. If time and status of
}
