% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_coxph_meta_analysis.R
\name{mi_coxph_meta_analysis}
\alias{mi_coxph_meta_analysis}
\title{COmparison of two treatments with cox regression}
\usage{
mi_coxph_meta_analysis(
  data,
  time,
  status,
  vars,
  var,
  meta.group,
  weights = FALSE,
  m = 25
)
}
\arguments{
\item{data}{data}

\item{time}{the time interval from start of observation until date of event (e.g. disease progression or death)
or censoring.}

\item{status}{variable specifying if event occured or data has been censored.}

\item{vars}{variables tested for Influence on outcome.}

\item{var}{grouping variable}

\item{meta.group}{variable for which meta analysis should be conducted. Usually the outcome of interest (e.g. treatment).}

\item{weights}{Logical value indicating wether inverse propensity score weighting should be applied}

\item{m}{number of iterations for multiple imputation}
}
\description{
This code generates the input for a forestplot comparing two treatment options.
}
