% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_PS.R
\name{mi_PS}
\alias{mi_PS}
\title{Calculate Propensity Scores with multiple imputation}
\usage{
mi_PS(data, vars, prop.var = NULL, m = 5, ...)
}
\arguments{
\item{data}{data.frame or data.table containing survival data.}

\item{vars}{variables tested for Influence on outcome.}

\item{prop.var}{variable for which propensity scores should be calculated. If no value is provided (prop.var = NULL), no weights are used in coxph. Default is NULL.}

\item{m}{Number of multiple imputations. The default is \code{m=5}.}

\item{...}{additional arguments to be passed on to coxph function}
}
\description{
This code generates propensity scores with multiple imputed values for missing data. Imputation is conducted with the \code{\link[mice:mice]{mice::mice()}} function.
Results are averaged for each row of entries representing single iterations.
}
