% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_list_mean.R
\name{nested_list_mean}
\alias{nested_list_mean}
\title{Calculate mean for each list element in a nested list}
\usage{
nested_list_mean(list, vars)
}
\arguments{
\item{list}{nested list}

\item{vars}{variables tested for Influence on outcome.}
}
\description{
This code generates takes each element i of nested list j and calculates the mean of the list elements with the same index.
}
