% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_table_survival.R
\name{outcome_table_survival}
\alias{outcome_table_survival}
\title{Generate outcome table}
\usage{
outcome_table_survival(
  data,
  timePFS = NULL,
  timeOS = NULL,
  statusPFS = NULL,
  statusOS = NULL,
  var,
  bestres = NULL,
  ORR = NULL,
  DCR = NULL,
  footnote,
  font = "calibri"
)
}
\arguments{
\item{data}{data.frame or data.table}

\item{timePFS}{Time for PFS}

\item{timeOS}{Time for OS}

\item{statusPFS}{censor variable for PFS}

\item{statusOS}{censor variable for OS}

\item{var}{Variable tested for Influence on outcome}

\item{bestres}{column containing data for best response}

\item{ORR}{column containing data for over all response rate}

\item{DCR}{column containing data for disease control rate}

\item{footnote}{add footnote}

\item{font}{font style for the table}
}
\description{
Wrapper around \code{add_median_survival} to transform the data to htmltable format in order to be used in display
with \code{htmlTable::htmlTable}. If time and status of
}
