% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_equation.R
\name{parse_equation}
\alias{parse_equation}
\title{Create and parse a model equation}
\usage{
parse_equation(equation, envir)
}
\arguments{
\item{equation}{A text defining the equation.}

\item{envir}{A named list with the indexes for the array.}
}
\value{
An expression with the code to evaluate the equation.
}
\description{
Create and parse a model equation
}
\examples{
equation <- list(
equation = "y[c] = x[c]^2",
indexes = "c in C"
)

C <- c("sec1", "sec2", "sec3")

parse_equation(equation, envir = globalenv())

}
