% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_chart_eumelareg.R
\name{pie_chart_eumelareg}
\alias{pie_chart_eumelareg}
\title{Pie Chart in EuMelaReg style}
\usage{
pie_chart_eumelareg(
  y,
  text.color = "white",
  text.size = 6,
  x.pos = 1.2,
  cols = NULL,
  title = NULL
)
}
\arguments{
\item{y}{Vector with group names (as character or factor)}

\item{text.color}{Color of the text within the chart to display percentages.}

\item{text.size}{Size of the text within the chart.}

\item{x.pos}{Position of text labels. Decrease to get closer to the origin and increase to move text towards the outer border.}

\item{cols}{Color scheme for the pie chart. Can be an \code{emr_cols} object or manual scale.}

\item{title}{Plot title.}
}
\description{
This code generates a pie chart to show distribution of variables.
}
