% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{require_columns}
\alias{require_columns}
\title{Check that a data.frame has required columns}
\usage{
require_columns(df, vec_cols)
}
\arguments{
\item{df}{data.frame}

\item{vec_cols}{vector of column names}
}
\value{
nothing, pops an error if any of the columns
        are not present in df
}
\description{
Check that a data.frame has required columns
}
