% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{roundUp}
\alias{roundUp}
\title{Part of helper functions used in this package}
\usage{
roundUp(x)
}
\arguments{
\item{x}{Vector of values to be rounded.}
}
\description{
This function rounds values to the closest number, which can be divided by five and adjust to different orders of
magnitude by using different exponentials of five depending on the difference between the maximum and minimum value in x.
}
