% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey_plot_eumelareg.R
\name{sankey_plot_eumelareg}
\alias{sankey_plot_eumelareg}
\title{An emR function}
\usage{
sankey_plot_eumelareg(data)
}
\arguments{
\item{data}{data.frame containing the different strata with one row for each patient}
}
\description{
This function
}
