% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_pie_chart_eumelareg.R
\name{save_pie_chart_eumelareg}
\alias{save_pie_chart_eumelareg}
\title{Save Pie Chart generated with \code{pie_chart_eumelareg}}
\usage{
save_pie_chart_eumelareg(
  data,
  y,
  regID,
  registry = "EUMELAREG",
  title = NULL,
  path = NULL,
  ...
)
}
\arguments{
\item{data}{dataframe.}

\item{y}{Name of column to be shown.}

\item{regID}{Name of column containing information about registries.}

\item{registry}{Name of EuMelaReg registry.}

\item{title}{Plot title.}

\item{path}{File path where the plot shall be stored. By default a folder "Results/pie_charts" is generated
within the working directory and objects are stored in this subfolder.}

\item{...}{Additional parameters passed on to \code{pie_chart_eumelareg}}
}
\description{
This code saves a pie chart to show distribution of variables grouped by EuMelaReg registries.
}
