% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_surv_plot_EMR.R
\name{save_surv_plot_EMR}
\alias{save_surv_plot_EMR}
\title{Save survival plot generated with survplot_eumelareg as png}
\usage{
save_surv_plot_EMR(
  data,
  time,
  status,
  var,
  weights = NULL,
  path,
  width = 12,
  height = 8,
  ...
)
}
\arguments{
\item{data}{data.frame or data.table containing survival data.}

\item{time}{the time interval from start of observation until date of event (e.g. disease progression or death)
or censoring.}

\item{status}{variable specifying if event occured or data has been censored.}

\item{var}{variable tested for Influence on outcome.}

\item{weights}{character variable specifying the name of the weights column. Weights have to be added to the original dataframe in order to be applied correctly.}

\item{path}{path where the files should be saved}

\item{width}{width of the png file in inches}

\item{height}{height of the png file in inches}

\item{...}{additional arguments to be passed on to survplot_eumelareg() function}
}
\description{
This function saves the output of  survplot_eumelareg as a png file with a default resolution of 600dpi in a 12x8 in format.
}
