% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_surv_plot_eumelareg.R
\name{save_surv_plot_eumelareg}
\alias{save_surv_plot_eumelareg}
\title{Save survival curve generated with \code{survplot_eumelareg}}
\usage{
save_surv_plot_eumelareg(
  data,
  LOT,
  survival = "PFS",
  var,
  regID = "REGISTR",
  registry = "EUMELAREG",
  path = NULL,
  ...
)
}
\arguments{
\item{data}{dataframe containing survival data.}

\item{LOT}{Line of treatment. Numerical value, either 1 or 2.}

\item{survival}{Specify which survival shall be plotted. Allowed values include "PFS" for progression free survival
and "OS" for overall survival. Default is "PFS".}

\item{var}{Grouping variable to compare survival.}

\item{regID}{Name of column containing registry information. Default is "REGISTR" as defined by SAS output.}

\item{registry}{Name of EuMelaReg registry. If \code{registry = "EUMELAREG"} data from all registries is plotted.}

\item{path}{File path where the plot shall be stored. By default a folder "Results/survival_curve" is generated
within the working directory and objects are stored in this subfolder.}

\item{...}{Additional arguments passed to \code{\link[survminer]{ggsurvplot}}.}
}
\description{
This function saves a default EuMelaReg survival plot (Kaplan-Meier plot) produced with \code{survminer}.
}
