% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_surv_plot__list_eumelareg.R
\name{save_surv_plot_list_eumelareg}
\alias{save_surv_plot_list_eumelareg}
\title{Save a list of plots generated with \code{save_surv_plot_eumelareg}}
\usage{
save_surv_plot_list_eumelareg(data, var1, var2 = NULL, ...)
}
\arguments{
\item{data}{dataframe containing survival data.}

\item{var1}{Grouping variable at first line of treatment.}

\item{var2}{If different value for second line of treatment (e.g. ECOG or LDH) please specify here.}

\item{...}{Additional arguments passed on to \code{\link[survminer]{ggsurvplot}}.
If NULL var1 is recycled (e.g. for GENDER).}
}
\description{
This function saves a list of default EuMelaReg survival plots (Kaplan-Meier plot) produced with \code{survminer}.
The default are 4 plots (1 for each Line of treatment and survival type pair).
}
