% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_armington.R
\name{simple_armington}
\alias{simple_armington}
\title{Create a simple armington model}
\usage{
simple_armington(v0, eps, eta, sigma, regions)
}
\arguments{
\item{v0}{a vector with the initial demand values.}

\item{eps}{a vector with supply elasticities.}

\item{eta}{the total demand price elasticity.}

\item{sigma}{the armington elasticity.}

\item{regions}{a vector with the regions names.}
}
\value{
Returns a list with the model components.
}
\description{
The model is writen using the "exact hat algebra".

For example, \eqn{\hat{q}_s = q_s'/q_s}, where q_s' is the value in the new equilibrium and q_s is the value in the old equilibrium.
}
\examples{

library(emr)

sa_model <- simple_armington(
 v0 = c(60, 30, 10),
 eps = c(1, 10, 10),
 eta = -1,
 sigma = 4,
 regions = c("dom", "sub", "nsub")
)

}
