% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_emr_block.R
\name{solve_emr_block}
\alias{solve_emr_block}
\title{Solve the model}
\usage{
solve_emr_block(model, start = NULL, ...)
}
\arguments{
\item{model}{a named list with params, variables and equations.}

\item{start}{a vector with start values for undefined variables. If NULL, the currente value will be used.}

\item{...}{the other arguments to be passed into the solver control argument.}

\item{method}{the solver to be used: BB or nleqslv}
}
\value{
Returns data and information about the model and its solution
}
\description{
Solve the model
}
\examples{

library(emr)

sa_model <- simple_armington(
v0 = c(60, 30, 10),
eps = c(1, 10, 10),
eta = -1,
sigma = 4,
regions = c("dom", "sub", "nsub")
)

# tau is the power tariff change
# Increase the tariff in 10 percent for the sub region
sa_model$params$tau$value['sub'] <- 1.1

sol <- solve_emr(sa_model)

sol$variables
}
