% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_cox_table.R
\name{strat_cox_table}
\alias{strat_cox_table}
\title{Print results from stratified univariate cox regression}
\usage{
strat_cox_table(
  data,
  time,
  status,
  vars,
  footnote = NULL,
  strata,
  printHTML = TRUE,
  univariate = TRUE,
  ...
)
}
\arguments{
\item{data}{data.frame or data.table containing survival data.}

\item{time}{the time interval from start of observation until date of event (e.g. disease progression or death)
or censoring.}

\item{status}{variable specifying if event occured or data has been censored.}

\item{vars}{one or more variables defined as character strings to be included in the table}

\item{footnote}{character string passed on to \code{tfoot} argument of \code{htmlTable}}

\item{strata}{factor to stratidfy by}

\item{printHTML}{Logical value. If TRUE output is printed as htmlTable. Default is TRUE.}

\item{univariate}{Logical value. If TRUE output of univariate cox regression is printed. Else output of multivariate
cox regression is printed. Not defined yet. TBD.}

\item{...}{additional arguments to be passed on to \code{cox_output}}
}
\description{
This function combines the results from multiple calls of \code{cox_output} and prints the output
as an htmlTable generated with the \code{htmlTable} package stratified by a factor.
}
