% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_time_table.R
\name{surv_time_table}
\alias{surv_time_table}
\title{Calculate x-months survival and show results in a table}
\usage{
surv_time_table(data, time, status, times, var)
}
\arguments{
\item{data}{data.frame or data.table containing survival data.}

\item{time}{The time interval from start of observation until date of event (e.g. disease progression or death)
or censoring.}

\item{status}{Variable specifying if event occured or data has been censored. Default behaviour inherited from the \code{surv_fit} function in the \code{survival} package, with 0
indicating censored data and 1 indicating event.}

\item{times}{Time in months for which survival should be calculated.}

\item{var}{Variable tested for Influence on outcome.}
}
\description{
This function calculates x-months survival grouped by an optional stratification factor and displays the results in a Table.
}
