% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_time.R
\name{survival_time}
\alias{survival_time}
\title{Calculate survival fraction at given timepoints}
\usage{
survival_time(data, time, status, var = NULL, times)
}
\arguments{
\item{data}{data.frame or data.table containing survival data.}

\item{time}{The time interval from start of observation until date of event (e.g. disease progression or death)
or censoring.}

\item{status}{Variable specifying if event occured or data has been censored. Default behaviour inherited from the \code{surv_fit} function in the \code{survival} package, with 0
indicating censored data and 1 indicating event.}

\item{var}{Variable tested for Influence on outcome.}

\item{times}{Time in months for which survival should be calculated.}
}
\description{
This function calculates x-months survival grouped by an optional stratification argument.
}
