% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimplot_eumelareg.R
\name{swimplot_eumelareg}
\alias{swimplot_eumelareg}
\title{Swimmer plot in EuMelaReg style}
\usage{
swimplot_eumelareg(
  data,
  ID,
  end,
  response.start,
  response.end,
  response.type,
  continued.response,
  durable.response,
  strat = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{data}{data.frame or data.table containing response data. Each event of response has to be defined in an own row.
Data for Response type, durable response, ongoing treatment and treatmend end have to entered in the respective columns.}

\item{ID}{column name that contains subject IDs.}

\item{end}{column name that contains time after which the subject died or was lost to follow-up.}

\item{response.start}{column that contains time after index date (e.g. diagnosis, treatment start) with documented response.}

\item{response.end}{column that contains time after index date with documented end of response.}

\item{response.type}{column that contains information whether patient had complete or partial response.}

\item{continued.response}{column that contains (logical) value whether response is ongoing or not.}

\item{durable.response}{column containing (logical) value whether response was durable (equal to or longer than six months).}

\item{strat}{grouping factor. Color of bars are colored accordingly.}

\item{xlab}{label of x-axis}

\item{ylab}{label of y-axis}
}
\description{
This function plots a default EuMelaReg swimmerplot using \code{ggplot}.
}
