% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_pval.R
\name{table1_pval}
\alias{table1_pval}
\title{Print table1 with pvalues}
\usage{
table1_pval(data, strat, vars, footnote = NULL, white.adjust = TRUE)
}
\arguments{
\item{data}{data.frame or data.table containing patient characteristics and demographics}

\item{strat}{strata to group by}

\item{vars}{character vector specifying columns that shall be included in the table}

\item{footnote}{A character string to be added as a footnote to the table.
Can also be a vector which results in multiple lines of footnotes.
The default \code{NULL} causes the footnote to be omitted.}

\item{white.adjust}{Default is FALSE. If TRUE, heteroscedasticity correction
is applied to the coefficient of covariance matrix. Used only for
independent measures ANOVA.}
}
\description{
This function calculates statistics and prints results in a table generated with \code{\link[=table1]{table1()}}. For categorical data
chi² test is used, for numerical data Welch ANOVA is conducted
}
