% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl1_to_flextable.R
\name{tbl1_to_flextable}
\alias{tbl1_to_flextable}
\title{Transform table1 to flextable object}
\usage{
tbl1_to_flextable(tbl1, pgwidth = 7, textsize = 9, headsize = 10)
}
\arguments{
\item{tbl1}{table1 object}

\item{pgwidth}{width of the table in word document}

\item{textsize}{fontsize of text body}

\item{headsize}{fontsize of header}
}
\description{
Transform table1 to flextable object
}
