% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_eumelareg_barplot.R
\name{theme_eumelareg_barplot}
\alias{theme_eumelareg_barplot}
\title{An emR function}
\usage{
theme_eumelareg_barplot(textsize = 14)
}
\arguments{
\item{textsize}{The default textsize used for Barplots}
}
\description{
This function defines the basic graphical representation of barplots for EuMelaReg
}
