% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_eumelareg_base.R
\name{theme_eumelareg_base}
\alias{theme_eumelareg_base}
\title{An emR function}
\usage{
theme_eumelareg_base()
}
\description{
This function defines the base theme for eumelareg functions with longer y-axis ticks and corbel as the default font.
}
