% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_eumelareg_pie_chart.R
\name{theme_eumelareg_pie_chart}
\alias{theme_eumelareg_pie_chart}
\title{An emR function}
\usage{
theme_eumelareg_pie_chart()
}
\description{
This function defines the basic graphical representation of barplots for EuMelaReg
}
