% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_eumelareg_surv_plot.R
\name{theme_eumelareg_surv_plot}
\alias{theme_eumelareg_surv_plot}
\title{An emR function}
\usage{
theme_eumelareg_surv_plot(textsize = 12)
}
\arguments{
\item{textsize}{The default textsize used for survival curves.}
}
\description{
This function defines the basic graphical representation of survival curves (Kaplan-Meier Plots) for EuMelaReg
}
