% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_eumelareg_surv_table.R
\name{theme_eumelareg_surv_table}
\alias{theme_eumelareg_surv_table}
\title{An emR function}
\usage{
theme_eumelareg_surv_table()
}
\description{
This function defines the basic graphical representation of the table below survival curves for EuMelaReg, showing
the number of patients at risk at a given time.
}
