% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_flextable.R
\name{theme_flextable}
\alias{theme_flextable}
\title{Define common formatting for flextables}
\usage{
theme_flextable(ft, textsize = 9, headsize = 10)
}
\arguments{
\item{ft}{flextable object}

\item{textsize}{fontsize of text body}

\item{headsize}{fontsize of header}
}
\description{
Define common formatting for flextables
}
