% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univ_cox_table.R
\name{univ_cox_table}
\alias{univ_cox_table}
\title{Combine results from multiple univariate cox regressions into one table}
\usage{
univ_cox_table(data, time, status, vars, footnote = NULL)
}
\arguments{
\item{data}{data.frame or data.table containing survival data.}

\item{time}{the time interval from start of observation until date of event (e.g. disease progression or death)
or censoring.}

\item{status}{variable specifying if event occured or data has been censored.}

\item{vars}{one or more variables defined as character strings to be included in the table}

\item{footnote}{character string passed on to \code{tfoot} argument of \code{htmlTable}}
}
\description{
This function combines the results from multiple calls of \code{univariate_cox_output} and prints the output
as an htmlTable generated with the \code{htmlTable} package.
}
