% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_cox_output.R
\name{univariate_cox_output}
\alias{univariate_cox_output}
\title{Calculate univariate cox regression and extract Hazard Ratio and pvalues}
\usage{
univariate_cox_output(data, time, status, var)
}
\arguments{
\item{data}{data.frame or data.table containing survival data.}

\item{time}{The time interval from start of observation until date of event (e.g. disease progression or death)
or censoring.}

\item{status}{Variable specifying if event occured or data has been censored.}

\item{var}{Variable tested for Influence on outcome.}
}
\description{
This function calculates a univaraite cox regression with the \code{coxph} function of the \code{survival}
package. Hazard ratios and pvalues are extracted and shown in a table format. If a variable is divided in more
than 2 groups the wald test statistic for overall significance is also shown.
}
