% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_median_survival.R
\name{add_median_survival}
\alias{add_median_survival}
\title{Calculate median survival time}
\usage{
add_median_survival(data, time, status, var, statistics = TRUE)
}
\arguments{
\item{data}{data.frame or data.table containing survival data.}

\item{time}{The time interval from start of observation until date of event (e.g. disease progression or death)
or censoring.}

\item{status}{Variable specifying if event occured or data has been censored. Default behaviour inherited from the \code{surv_fit} function in the \code{survival} package, with 0
indicating censored data and 1 indicating event.}

\item{var}{Variable tested for Influence on outcome.}
}
\description{
Fitting a survival fit for each subgroup defined by var. Additionally the total survival time for the
whole sample is calculated.
}
