% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_eumelareg.R
\name{barplot_eumelareg}
\alias{barplot_eumelareg}
\title{An emR function}
\usage{
barplot_eumelareg(
  data,
  x,
  y,
  n.vjust = 0.05,
  bar.border.size = 1,
  bar.alpha = 0.5,
  bar.width = 0.5,
  y.breaks = breakFun,
  errorbar.fun = "mean_sdl",
  errorbar.width = 0.3,
  errorbar.size = 1.3,
  textsize = 14,
  brewer.palette = "RdYlBu",
  ylab = "Value",
  scale = 5,
  dotplot.binwidth = 1
)
}
\arguments{
\item{data}{data.frame or data.table.}

\item{x}{Independent variable}

\item{y}{Dependent variable}

\item{n.vjust}{Offset of n over the maximum value for each group}

\item{bar.border.size}{Size of the border surrounding the bars.}

\item{bar.alpha}{Alpha value of the barplot fill argument. If < 1 a semi-transparent plot is generated and
the graph cannot be saved as metafile anymore.}

\item{bar.width}{Width of the bars.}

\item{y.breaks}{Argument to define breaks on y-axis. Default is \code{breakFun}, which displays the breaks as
a multiple of 5.}

\item{errorbar.fun}{Function used for calculation of the errrorbars. Default is standard deviation.}

\item{errorbar.width}{Width of the errorbar whiskers.}

\item{errorbar.size}{Thickness of errorbar line.}

\item{brewer.palette}{Color palette from package \code{RColorBrewer} or custom color palette.}

\item{ylab}{Label for y-axis.}

\item{dotplot.binwidth}{Indirectly regulates size of the dots. Relative to the plot size.}
}
\description{
This function plots a default EuMelaReg Barplot produced with \code{ggplot2}. Similar to a ggplot2 object,
additional custom layers can be added with the \code{+} operator.
}
\examples{
dt <- data.table(group = factor(rep(1:6,500)), value = rnorm(3000, 50,15))
barplot_eumelareg(data = dt, x="group", y = "value", n.vjust =.07, 
                  dotplot.binwidth = 0.5, brewer.palette = "Blues")
}
