% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sankey_flow.R
\name{calc_sankey_flow}
\alias{calc_sankey_flow}
\title{Calculate data flow}
\usage{
calc_sankey_flow(x, y, data)
}
\arguments{
\item{x}{number of column containing source data}

\item{y}{number of column containing target data}

\item{data}{data.frame containing the different strata with one row for each patient}
}
\description{
This function calculates the flow from 1 stratum to the next.
}
