% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_output.R
\name{cox_output}
\alias{cox_output}
\title{Calculate cox regression and extract Hazard Ratio and pvalues}
\usage{
cox_output(
  data,
  time,
  status,
  vars,
  fixed.var = NULL,
  modeltype = "full",
  p.thres = 0.1,
  niter = 10
)
}
\arguments{
\item{data}{data.frame or data.table containing survival data.}

\item{time}{the time interval from start of observation until date of event (e.g. disease progression or death)
or censoring.}

\item{status}{variable specifying if event occured or data has been censored.}

\item{vars}{variables tested for Influence on outcome. Defines type of cox regression: if length is 1 the output
is a univariate cox regression, for length > 1 the output is a multivariate cox regression}

\item{fixed.var}{specifies fixed variables to be included in the cox model.}

\item{modeltype}{character value. Allowed values include: "full" and "backwards".
Specify if full model or backwards selection model should be used for multivariate cox regression.}

\item{p.thres}{pvalue threshold for backwards selection model.}

\item{niter}{number of iterations for backwards selection model.}
}
\description{
This function calculates a univariate cox regression with the \code{coxph} function from the \code{survival}
package if the argument vars is length 1; otherwise a multivariate cox regression is calculated. Hazard ratios and pvalues are extracted and shown in a table format.
If a variable in univariate cox regression is divided in more than 2 groups the wald test statistic
for overall significance is also shown.
}
