% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_meta_analysis.R
\name{coxph_meta_analysis}
\alias{coxph_meta_analysis}
\title{Meta analysis forest plot}
\usage{
coxph_meta_analysis(
  data,
  time,
  status,
  vars,
  var,
  meta.group,
  weights = NULL,
  univariate = FALSE,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{data}{data}

\item{time}{the time interval from start of observation until date of event (e.g. disease progression or death)
or censoring.}

\item{status}{variable specifying if event occured or data has been censored.}

\item{vars}{variables tested for Influence on outcome.}

\item{var}{grouping variable}

\item{meta.group}{variable for which meta analysis should be conducted. Usually the outcome of interest (e.g. treatment).}

\item{weights}{character variable specifying the name of the weights column. Weights have to be added to the original dataframe in order to be applied correctly.}

\item{univariate}{Logical value. If TRUE output of univariate cox regression is printed. Else output of multivariate
cox regression is printed. Default is FALSE.}

\item{na.rm}{logical indicating whether missing values (Specified as "missing" or "Missing") should be removed}

\item{...}{additional arguments passed on to coxph}
}
\description{
This code generates a forestplot from a meta analysis coxph model.
}
