% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot_eumelareg.R
\name{forestplot_eumelareg}
\alias{forestplot_eumelareg}
\title{Forest Plot for Cox Proportional Hazards Model in EuMelaReg style}
\usage{
forestplot_eumelareg(
  model,
  data = NULL,
  vars = NULL,
  main = "Hazard ratio for disease progression or death (95\% CI)",
  y_breaks = NULL,
  cpositions = c(0.02, 0.22, 0.4),
  point.size = 3,
  fontsize = 0.7,
  line.size = 0.7,
  vjust_text = 1.2,
  refLabel = "reference",
  noDigits = 2,
  varnames = NULL
)
}
\arguments{
\item{model}{an object of class coxph.}

\item{data}{a dataset used to fit survival curves. If not supplied then data
will be extracted from 'fit' object.}

\item{vars}{variables that were used in the coxph model}

\item{main}{title of the plot.}

\item{y_breaks}{argument to supply manual y_breaks as a numerical vector. Default is NULL and breaks are set automatically within the function.}

\item{cpositions}{relative positions of first three columns in the OX scale.}

\item{point.size}{Size of mean points.}

\item{fontsize}{relative size of annotations in the plot. Default value: 0.7.}

\item{line.size}{Size of errorbar line.}

\item{vjust_text}{vertical adjustment of text containing information about events, global pvalue, AIC and concordance index}

\item{refLabel}{label for reference levels of factor variables.}

\item{noDigits}{number of digits for estimates and p-values in the plot.}

\item{varnames}{Character vector specifying rownames of the table (empty columns should be named with "").}
}
\description{
This code generates a forest plot from a coxph model.
}
