% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot_eumelareg.R
\name{forestplot_eumelareg}
\alias{forestplot_eumelareg}
\title{An emR function}
\usage{
forestplot_eumelareg(
  model,
  data = NULL,
  main = "Hazard ratio for disease progression or death (95\% CI)",
  cpositions = c(0.02, 0.22, 0.4),
  fontsize = 0.7,
  refLabel = "reference",
  noDigits = 2,
  varnames = NULL
)
}
\arguments{
\item{model}{an object of class coxph.}

\item{data}{a dataset used to fit survival curves. If not supplied then data
will be extracted from 'fit' object.}

\item{main}{title of the plot.}

\item{cpositions}{relative positions of first three columns in the OX scale.}

\item{fontsize}{relative size of annotations in the plot. Default value: 0.7.}

\item{refLabel}{label for reference levels of factor variables.}

\item{noDigits}{number of digits for estimates and p-values in the plot.}

\item{varnames}{Character vector specifying rownames of the table (empty columns should be named with "").}
}
\description{
This code generates a pie chart to show distribution of variables.
}
