% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot_meta_eumelareg.R
\name{forestplot_meta_eumelareg}
\alias{forestplot_meta_eumelareg}
\title{Meta analysis forest plot}
\usage{
forestplot_meta_eumelareg(
  data,
  time,
  status,
  vars,
  meta.group,
  univariate = TRUE,
  weights = FALSE,
  imputation = FALSE,
  main = "Hazard ratio for disease progression or death (95\% CI)",
  y_breaks = NULL,
  cpositions = c(0, 0.1, 0.3),
  impIter = 25,
  refLabel = "reference",
  point_size = 4,
  fontsize = 1,
  line_size = 0.9,
  vjust_text = 1.2,
  noDigits = 2,
  varnames = NULL,
  ylim = NULL
)
}
\arguments{
\item{data}{data}

\item{time}{the time interval from start of observation until date of event (e.g. disease progression or death)
or censoring.}

\item{status}{variable specifying if event occured or data has been censored.}

\item{vars}{variables tested for Influence on outcome.}

\item{meta.group}{variable for which meta analysis should be conducted. Usually the outcome of interest (e.g. treatment).}

\item{univariate}{Logical value. If TRUE output of univariate cox regression is printed. Else output of multivariate}

\item{weights}{Logical variable specifying whether inverse propensity score weighting should be applied.}

\item{imputation}{Logical variable specifying whether missing should be replaced by multiple imputation}

\item{main}{title of the plot.}

\item{y_breaks}{argument to supply manual y_breaks as a numerical vector. Default is NULL and breaks are set automatically within the function.}

\item{cpositions}{relative positions of first three columns in the OX scale.}

\item{impIter}{number of iterations for multiple imputation.}

\item{refLabel}{label for reference levels of factor variables.}

\item{point_size}{Size of mean points.}

\item{fontsize}{relative size of annotations in the plot. Default value: 0.7.}

\item{line_size}{Size of errorbar line.}

\item{vjust_text}{vertical adjustment of text containing information about events, global pvalue, AIC and concordance index}

\item{noDigits}{number of digits for estimates and p-values in the plot.}

\item{varnames}{Character vector specifying rownames of the table (empty columns should be named with "").}

\item{ylim}{argument to supply manual y limits as numerical vector of length 2. Default is NULL and limits are set automatically within the function.}
}
\description{
This code generates a forestplot from a meta analysis coxph model.
}
