% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_table_survival.R
\name{outcome_table_survival}
\alias{outcome_table_survival}
\title{Generate outcome table}
\usage{
outcome_table_survival(
  data,
  time,
  status,
  surv_names,
  var,
  bestres = NULL,
  ORR = NULL,
  DCR = NULL,
  statistics = TRUE,
  footnote = NULL,
  font = "calibri",
  ...
)
}
\arguments{
\item{data}{data.frame or data.table}

\item{time}{Time for survival}

\item{status}{censor variable}

\item{surv_names}{names of the survival to be displayed in the table (e.g "Median PFS (95\% CI)")}

\item{var}{Variable tested for Influence on outcome}

\item{bestres}{column containing data for best response}

\item{ORR}{column containing data for over all response rate}

\item{DCR}{column containing data for disease control rate}

\item{statistics}{Logical value. If TRUE pvalue is printed. Default is TRUE. Default test statistics are wilcoxon (or anova if n > 2)}

\item{footnote}{add footnote}

\item{font}{font style for the table}

\item{...}{add additional css styling arguments to addHtmlTableStyle from the \code{htmlTable} package
for numerical data and fisher exact test for categorical data.}
}
\description{
Wrapper around \code{add_median_survival} to transform the data to htmltable format in order to be used in display
with \code{htmlTable::htmlTable}. If time and status of
}
