% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_surv_plot_eumelareg.R
\name{save_surv_plot_eumelareg}
\alias{save_surv_plot_eumelareg}
\title{Save survival curve generated with \code{survplot_eumelareg}}
\usage{
save_surv_plot_eumelareg(
  data,
  time,
  status,
  var,
  legend.labs = NULL,
  regID = "REGISTR",
  registry = "EUMELAREG",
  path = NULL,
  ...
)
}
\arguments{
\item{data}{dataframe containing survival data.}

\item{time}{The time interval from start of observation until date of event (e.g. disease progression or death)
or censoring.}

\item{status}{Variable specifying if event occured or data has been censored.}

\item{var}{Variable tested for Influence on outcome.}

\item{legend.labs}{Character vector specifying legend labels. Used to replace the names of the strata from the fit.}

\item{regID}{Name of column containing registry information. Default is "REGISTR" as defined by SAS output.}

\item{registry}{Name of EuMelaReg registry. If \code{registry = "EUMELAREG"} data from all registries is plotted.}

\item{path}{File path where the plot shall be stored. By default a folder "Results/survival_curve" is generated
within the working directory and objects are stored in this subfolder.}

\item{...}{Additional arguments passed to \code{\link[survminer]{ggsurvplot}}.}
}
\description{
This function saves a default EuMelaReg survival plot (Kaplan-Meier plot) produced with \code{survminer}.
}
