% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survplot_eumelareg.R,
%   R/survplot_eumelareg_boot.R
\name{survplot_eumelareg}
\alias{survplot_eumelareg}
\title{Survival curve in EuMelaReg style}
\usage{
survplot_eumelareg(
  data,
  time = "time",
  status = "status",
  var = NULL,
  xlab = "Time in months",
  ylab = "Probability of Survival",
  axes.offset = FALSE,
  break.y.by = 0.1,
  break.time.by = 3,
  xlim = c(0, 48),
  ggtheme = theme_eumelareg_surv_plot(),
  tables.theme = theme_eumelareg_surv_table(),
  plot.width = 0.838,
  plot.height = 0.88,
  plot.margin.left = NULL,
  text.size = 12,
  weights = NULL,
  landmarks = c(12, 24),
  conf.type = "log-log",
  risk.table.width = 0.9,
  risk.table.title = NULL,
  legend.position = "top",
  legend.title = "",
  legend.labs = NULL,
  pval = TRUE,
  pval.coord = c(1, 0.1),
  merge = FALSE,
  palette = "jco",
  extract.legend = FALSE,
  ...
)

survplot_eumelareg(
  data,
  time = "time",
  status = "status",
  var = NULL,
  xlab = "Time in months",
  ylab = "Probability of Survival",
  axes.offset = FALSE,
  break.y.by = 0.1,
  break.time.by = 3,
  xlim = c(0, 48),
  ggtheme = theme_eumelareg_surv_plot(),
  tables.theme = theme_eumelareg_surv_table(),
  plot.width = 0.838,
  plot.height = 0.88,
  plot.margin.left = NULL,
  text.size = 12,
  weights = NULL,
  landmarks = c(12, 24),
  conf.type = "log-log",
  risk.table.width = 0.9,
  risk.table.title = NULL,
  legend.position = "top",
  legend.title = "",
  legend.labs = NULL,
  pval = TRUE,
  pval.coord = c(1, 0.1),
  merge = FALSE,
  palette = "jco",
  extract.legend = FALSE,
  ...
)
}
\arguments{
\item{data}{data.frame or data.table containing survival data.}

\item{time}{the time interval from start of observation until date of event (e.g. disease progression or death)
or censoring.}

\item{status}{variable specifying if event occured or data has been censored.}

\item{var}{variable tested for Influence on outcome.}

\item{xlab}{X-axis label.}

\item{ylab}{y-axis label.}

\item{axes.offset}{logical value. If TRUE the space between the plot origin and the axes is removed.}

\item{break.y.by}{interval of breaks on the y-axis.}

\item{break.time.by}{Interval of breaks on the x-axis (time axis).}

\item{xlim}{x axis limit.}

\item{ggtheme}{function, ggplot2 theme name. Default value is theme_eumelareg_surv_plot. Allowed values include ggplot2 official themes: see \code{theme}.}

\item{tables.theme}{function, ggplot2 theme name. Default value is theme_eumelareg_surv_plot. Allowed values include ggplot2 official themes: see \code{theme}.}

\item{plot.width}{relative width of the survival plot}

\item{plot.height}{relative height of the survival plot. The risk table is adjusted accordingly.}

\item{plot.margin.left}{numerical. Used to adjust the plot horizontally.}

\item{text.size}{size of plot text.}

\item{weights}{character variable specifying the name of the weights column. Weights have to be added to the original dataframe in order to be applied correctly.}

\item{landmarks}{times for which landmark survival data shall be provided.
Should be given in the same order as those strata.}

\item{conf.type}{Method to calculate confidence intervals. Log-log method is the default in SAS.}

\item{risk.table.width}{relative width of the risk table.}

\item{risk.table.title}{the title to be used for the risk table. Default is no title.}

\item{legend.position}{he position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{legend.title}{name of legend title.}

\item{legend.labs}{character vector specifying legend labels. Used to replace the names of the strata from the fit.
Should be given in the same order as those strata.}

\item{pval}{logical value, a numeric or a string. If logical and TRUE, the
p-value is added on the plot. If numeric, than the computet p-value is
substituted with the one passed with this parameter. If character, then the
customized string appears on the plot. See examples - Example 3.}

\item{pval.coord}{coords of pvalue within plot.}

\item{merge}{logical value. If TRUE survival curve and median survival table are plotted in the same graph. Else
two separate figures are generated. Default is FALSE.}

\item{palette}{the color palette to be used. Allowed values include "hue" for
the default hue color scale; "grey" for grey color palettes; brewer palettes
e.g. "RdBu", "Blues", ...; or custom color palette e.g. c("blue", "red"); and scientific journal palettes from ggsci R package, e.g.: "npg",
 "aaas", "lancet", "jco", "ucscgb", "uchicago", "simpsons" and
 "rickandmorty".
See details section for more information.  Can be also a numeric vector of
length(groups); in this case a basic color palette is created using the
function \link[grDevices]{palette}.}

\item{extract.legend}{logical indicating whether only the legend should be extracted from the plot to build manual arrangements of plots and legend with \code{\link[=ggarrange]{ggarrange()}}}

\item{...}{Futher arguments as described hereafter and
other arguments to be passed i) to ggplot2 geom_*() functions such
 as linetype, size, ii) or to the function \link[ggpubr]{ggpar}() for
 customizing the plots. See details section.}
}
\description{
This function plots a default EuMelaReg survival plot (Kaplan-Meier plot) produced with \code{survminer}.

This function plots a default EuMelaReg survival plot (Kaplan-Meier plot) produced with \code{survminer}.
}
\details{
Further arguments can be obtained from the \code{\link[=ggsurvplot]{ggsurvplot()}} function.

Further arguments can be obtained from the \code{\link[=ggsurvplot]{ggsurvplot()}} function.
}
