% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survplot_eumelareg.R
\name{survplot_eumelareg}
\alias{survplot_eumelareg}
\title{Survival curve in EuMelaReg style}
\usage{
survplot_eumelareg(
  data,
  time = "time",
  status = "status",
  var,
  xlab = "Time in months",
  ylab = "Probability of Overall Survival",
  pval = TRUE,
  break.y.by = 0.1,
  break.time.by = 3,
  ggtheme = theme_eumelareg_surv_plot(),
  tables.theme = theme_eumelareg_surv_table(),
  axes.offset = TRUE,
  risk.table = "absolute",
  risk.table.y.text = TRUE,
  risk.table.title = "No. at Risk",
  table.margin.left = 0,
  legend.labs = NULL,
  palette = "jco",
  pval.coord = c(0.05, 0.03),
  ...
)
}
\arguments{
\item{data}{data.frame or data.table containing survival data.}

\item{time}{The time interval from start of observation until date of event (e.g. disease progression or death)
or censoring.}

\item{status}{Variable specifying if event occured or data has been censored.}

\item{var}{Variable tested for Influence on outcome.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{pval}{logical value, a numeric or a string. If logical and TRUE, the
p-value is added on the plot. If numeric, than the computet p-value is
substituted with the one passed with this parameter. If character, then the
customized string appears on the plot. See examples - Example 3.}

\item{break.y.by}{Interval of breaks on the y-axis.}

\item{break.time.by}{Interval of breaks on the x-axis (time axis).}

\item{ggtheme}{function, ggplot2 theme name. Default value is theme_eumelareg_surv_plot. Allowed values include ggplot2 official themes: see \code{theme}.}

\item{tables.theme}{function, ggplot2 theme name. Default value is theme_eumelareg_surv_plot. Allowed values include ggplot2 official themes: see \code{theme}.}

\item{axes.offset}{Logical value. If TRUE the space between the plot origin and the axes is removed.}

\item{risk.table}{Allowed values include: \itemize{ \item TRUE or FALSE
specifying whether to show or not the risk table. Default is FALSE. \item
"absolute" or "percentage". Shows the \bold{absolute number} and the
\bold{percentage} of subjects at risk by time, respectively. \item "abs_pct"
to show both absolute number and percentage. \item "nrisk_cumcensor" and
"nrisk_cumevents". Show the number at risk and, the cumulative number of
censoring and events, respectively. }}

\item{risk.table.y.text}{Logical value. Default value is TRUE. If FALSE, the y axis tick labels of tables will be hidden.}

\item{risk.table.title}{The title to be used for the risk table}

\item{table.margin.left}{Numerical. Used to adjust the risk table horizontally.}

\item{legend.labs}{Character vector specifying legend labels. Used to replace the names of the strata from the fit.
Should be given in the same order as those strata.}

\item{palette}{the color palette to be used. Allowed values include "hue" for
the default hue color scale; "grey" for grey color palettes; brewer palettes
e.g. "RdBu", "Blues", ...; or custom color palette e.g. c("blue", "red"); and scientific journal palettes from ggsci R package, e.g.: "npg",
 "aaas", "lancet", "jco", "ucscgb", "uchicago", "simpsons" and
 "rickandmorty".
See details section for more information.  Can be also a numeric vector of
length(groups); in this case a basic color palette is created using the
function \link[grDevices]{palette}.}

\item{pval.coord}{Coords of pvalue within plot.}

\item{...}{Futher arguments as described hereafter and
other arguments to be passed i) to ggplot2 geom_*() functions such
 as linetype, size, ii) or to the function \link[ggpubr]{ggpar}() for
 customizing the plots. See details section.}
}
\description{
This function plots a default EuMelaReg survival plot (Kaplan-Meier plot) produced with \code{survminer}.
}
\seealso{
\code{\link[=ggsurvplot]{ggsurvplot()}}
}
